+++ src/andromede/libs/standard_sc.py
@@ -11,8 +11,8 @@
 # This file is part of the Antares project.
 
 
-from andromede.expression.expression_efficient import literal, param
-from andromede.expression.linear_expression_efficient import port_field, var
+from andromede.expression.expression import literal, param
+from andromede.expression.linear_expression import port_field, var
 from andromede.libs.standard import BALANCE_PORT_TYPE, CONSTANT
 from andromede.model import (
     Constraint,
