+++ src/andromede/model/common.py
@@ -16,8 +16,8 @@ Module for common classes used in models.
 from enum import Enum
 from typing import Union
 
-from andromede.expression.expression_efficient import ExpressionNodeEfficient
-from andromede.expression.linear_expression_efficient import LinearExpressionEfficient
+from andromede.expression.expression import ExpressionNodeEfficient
+from andromede.expression.linear_expression import LinearExpressionEfficient
 
 ValueOrExprNodeOrLinearExpr = Union[
     int, float, ExpressionNodeEfficient, LinearExpressionEfficient
