+++ tests/integration/test_benders_decomposed.py
@@ -13,9 +13,9 @@
 import pandas as pd
 import pytest
 
-from andromede.expression.expression_efficient import literal, param
+from andromede.expression.expression import literal, param
 from andromede.expression.indexing_structure import IndexingStructure
-from andromede.expression.linear_expression_efficient import var
+from andromede.expression.linear_expression import var
 from andromede.libs.standard import (
     BALANCE_PORT_TYPE,
     CONSTANT,
