+++ tests/models/test_electrolyzer.py
@@ -10,8 +10,8 @@
 #
 # This file is part of the Antares project.
 
-from andromede.expression.expression_efficient import literal, param
-from andromede.expression.linear_expression_efficient import port_field, var
+from andromede.expression.expression import literal, param
+from andromede.expression.linear_expression import port_field, var
 from andromede.libs.standard import CONSTANT, TIME_AND_SCENARIO_FREE
 from andromede.model import (
     Constraint,
